current_user/likesとcurrent_user/postsのAPIをAllモジュールに切り出した。
